Try to fix a warning on MSVC
```
...\include\CGAL/Polyhedral_mesh_domain_with_features_3.h(264): warning C4267: 'argument': conversion from 'size_t' to '_Ty', possible loss of data
        with
        [
            _Ty=int
        ]
...\cmake\platforms\x64_Cygwin-Windows10_MSVC2017-Release-64bits\test\Mesh_3\test_meshing_polylines_only.cpp(80): note: see reference to function template instantiation 'void CGAL::Polyhedral_mesh_domain_with_features_3<K,CGAL::Polyhedron_3<Gt,CGAL::Mesh_3::Mesh_polyhedron_items<Patch_id>,CGAL::HalfedgeDS_default,std::allocator<_Ty>>,CGAL::Default,Patch_id,CGAL::Tag_true>::add_features<std::_Vector_iterator<std::_Vector_val<std::_Simple_types<std::vector<CGAL::Point_3<Kernel_>,std::allocator<CGAL::Point_3<Kernel_>>>>>>>(InputIterator,InputIterator)' being compiled
        with
        [
            Gt=K,
            Patch_id=int,
            _Ty=int,
            Kernel_=CGAL::Epick,
            InputIterator=std::_Vector_iterator<std::_Vector_val<std::_Simple_types<Polyline>>>
        ]
```

+++ Mesh_3/include/CGAL/Polyhedral_mesh_domain_with_features_3.h
@@ -261,7 +261,7 @@ public:
         i != patch_ids_bitset.npos;
         i = patch_ids_bitset.find_next(i))
     {
-      all_patch_ids.push_back(i + min);
+      all_patch_ids.push_back(static_cast<int>(i + min));
     }
     using Polyline = typename std::iterator_traits<InputIterator>::value_type;
     auto identity_property_map = boost::typed_identity_property_map<Polyline>();
